feat: Refactor DataTable component and add SortableDataTableRow
