+++ src/components/common/DataTable/SortableDataTableRow/index.ts
@@ -1,0 +1,1 @@
+export { SortableDataTableRow } from "./SortableDataTableRow";
